+++ skoolkit/config.py
@@ -130,6 +130,7 @@ COMMANDS = {
     'trace': {
         'PNGScale': (2, ''),
         'ScreenFps': (50, ''),
+        'ScreenScale': (2, ''),
         'TraceLine': ("${pc:04X} {i}", ''),
         'TraceLine2': (
             "${pc:04X} {i:<15}  "