+++ tests/test_trace.py
@@ -1698,6 +1698,7 @@ class TraceTest(SkoolKitTestCase):
             "[trace]\n"
             "PNGScale=2\n"
             "ScreenFps=50\n"
+            "ScreenScale=2\n"
             "TraceLine=${pc:04X} {i}\n"
             "TraceLine2=${pc:04X} {i:<15}  "
             "A={r[a]:02X}  F={r[f]:08b}  BC={r[bc]:04X}  DE={r[de]:04X}  HL={r[hl]:04X}  IX={r[ix]:04X} IY={r[iy]:04X}\\n                       "
@@ -1724,6 +1725,7 @@ class TraceTest(SkoolKitTestCase):
             [trace]
             PNGScale=2
             ScreenFps=50
+            ScreenScale=2
             TraceLine=${pc:04X} {i}
             TraceLine2=${pc:04x} {i:<15} A={r[a]:02X}
             TraceLineDecimal={pc:05} {i}
@@ -2161,6 +2163,65 @@ class TraceTest(SkoolKitTestCase):
         self.assertEqual(screen.fps, 150)
         self.assertEqual(screen.caption, 'trace.py')
         self.assertEqual(screen.scr[0], 1)
+        self.assertEqual(screen.frame, 1)
+
+    @patch.object(trace, 'pygame', True)
+    @patch.object(trace, 'Screen', MockScreen)
+    def test_config_ScreenScale_read_from_file(self):
+        ini = """
+            [trace]
+            ScreenScale=1
+        """
+        self.write_text_file(dedent(ini).strip(), 'skoolkit.ini')
+        data = (
+            0x00, # $8000 NOP ; T=69886 (screen draw follows)
+            0x00, # $8001 NOP ; T=69890
+        )
+        start = 0x8000
+        stop = start + len(data)
+        ram = [0] * 49152
+        ram[0] = 240
+        ram[start - 0x4000:stop - 0x4000] = data
+        registers = {'PC': start, 'tstates': 69886}
+        z80file = self.write_z80_file(None, ram, registers=registers)
+        output, error = self.run_trace(f'-S {stop} --screen {z80file}')
+        self.assertEqual(error, '')
+        exp_output = f"""
+            Using pygame
+            Stopped at ${stop:04X}
+        """
+        self.assertEqual(dedent(exp_output).strip(), output.rstrip())
+        self.assertEqual(screen.scale, 1)
+        self.assertEqual(screen.fps, 50)
+        self.assertEqual(screen.caption, 'trace.py')
+        self.assertEqual(screen.scr[0], 240)
+        self.assertEqual(screen.frame, 1)
+
+    @patch.object(trace, 'pygame', True)
+    @patch.object(trace, 'Screen', MockScreen)
+    def test_config_ScreenScale_set_on_command_line(self):
+        data = (
+            0x00, # $8000 NOP ; T=69886 (screen draw follows)
+            0x00, # $8001 NOP ; T=69890
+        )
+        start = 0x8000
+        stop = start + len(data)
+        ram = [0] * 49152
+        ram[0] = 16
+        ram[start - 0x4000:stop - 0x4000] = data
+        registers = {'PC': start, 'tstates': 69886}
+        z80file = self.write_z80_file(None, ram, registers=registers)
+        output, error = self.run_trace(f'-I ScreenScale=3 -S {stop} --screen {z80file}')
+        self.assertEqual(error, '')
+        exp_output = f"""
+            Using pygame
+            Stopped at ${stop:04X}
+        """
+        self.assertEqual(dedent(exp_output).strip(), output.rstrip())
+        self.assertEqual(screen.scale, 3)
+        self.assertEqual(screen.fps, 50)
+        self.assertEqual(screen.caption, 'trace.py')
+        self.assertEqual(screen.scr[0], 16)
         self.assertEqual(screen.frame, 1)
 
     def test_config_TraceLine_read_from_file(self):